# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/products": {
      "post": {
        "operationId": "ProductsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Product created successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new product (admin only)",
        "tags": [
          "products"
        ]
      },
      "get": {
        "operationId": "ProductsController_findAll",
        "parameters": [
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search term for product name or description",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "categoryId",
            "required": false,
            "in": "query",
            "description": "Category ID to filter products",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "minPrice",
            "required": false,
            "in": "query",
            "description": "Minimum price",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "maxPrice",
            "required": false,
            "in": "query",
            "description": "Maximum price",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "tag",
            "required": false,
            "in": "query",
            "description": "Tag to filter products",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "featured",
            "required": false,
            "in": "query",
            "description": "Only include featured products",
            "schema": {
              "default": false,
              "type": "boolean"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Field to sort by",
            "schema": {
              "default": "createdAt",
              "type": "string",
              "enum": [
                "price",
                "name",
                "createdAt",
                "averageRating"
              ]
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "description": "Sort order",
            "schema": {
              "default": "DESC",
              "type": "string",
              "enum": [
                "ASC",
                "DESC"
              ]
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number for pagination",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "default": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return filtered products"
          }
        },
        "summary": "Get all products with filtering, sorting, and pagination",
        "tags": [
          "products"
        ]
      }
    },
...
    "/api/products/{id}": {
      "get": {
        "operationId": "ProductsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return the product"
          },
          "404": {
            "description": "Product not found"
          }
        },
        "summary": "Get a product by ID",
        "tags": [
          "products"
        ]
      },
      "put": {
        "operationId": "ProductsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Product updated successfully"
          },
          "404": {
            "description": "Product not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a product (admin only)",
        "tags": [
          "products"
        ]
      },
      "delete": {
        "operationId": "ProductsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product deleted successfully"
          },
          "404": {
            "description": "Product not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a product (admin only)",
        "tags": [
          "products"
        ]
      }
    },
...
  },
  "info": {
    "title": "Đông Y Pharmacy API",
    "description": "The Đông Y Pharmacy REST API documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateProductDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the product"
          },
          "description": {
            "type": "string",
            "description": "Detailed description of the product"
          },
          "price": {
            "type": "number",
            "description": "Regular price of the product"
          },
          "salePrice": {
            "type": "number",
            "description": "Sale price of the product"
          },
          "stockQuantity": {
            "type": "number",
            "description": "Stock quantity of the product"
          },
          "imageUrl": {
            "type": "string",
            "description": "URL to the main product image"
          },
          "additionalImages": {
            "description": "Additional product images",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "tags": {
            "description": "Tags associated with the product",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "isActive": {
            "type": "boolean",
            "description": "Whether the product is active",
            "default": true
          },
          "isFeatured": {
            "type": "boolean",
            "description": "Whether the product is featured",
            "default": false
          },
          "specifications": {
            "type": "string",
            "description": "Product specifications in detail"
          },
          "usageInstructions": {
            "type": "string",
            "description": "Usage instructions for the product"
          },
          "sku": {
            "type": "string",
            "description": "Stock keeping unit (SKU)"
          },
          "categoryId": {
            "type": "string",
            "description": "Category ID the product belongs to"
          }
        },
        "required": [
          "name",
          "description",
          "price",
          "stockQuantity",
          "categoryId"
        ]
      },
      "UpdateProductDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the product"
          },
          "description": {
            "type": "string",
            "description": "Detailed description of the product"
          },
          "price": {
            "type": "number",
            "description": "Regular price of the product"
          },
          "salePrice": {
            "type": "number",
            "description": "Sale price of the product"
          },
          "stockQuantity": {
            "type": "number",
            "description": "Stock quantity of the product"
          },
          "imageUrl": {
            "type": "string",
            "description": "URL to the main product image"
          },
          "additionalImages": {
            "description": "Additional product images",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "tags": {
            "description": "Tags associated with the product",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "isActive": {
            "type": "boolean",
            "description": "Whether the product is active"
          },
          "isFeatured": {
            "type": "boolean",
            "description": "Whether the product is featured"
          },
          "specifications": {
            "type": "string",
            "description": "Product specifications in detail"
          },
          "usageInstructions": {
            "type": "string",
            "description": "Usage instructions for the product"
          },
          "sku": {
            "type": "string",
            "description": "Stock keeping unit (SKU)"
          },
          "categoryId": {
            "type": "string",
            "description": "Category ID the product belongs to"
          }
        }
      },
...
    }
  }
}